Polish UI pt 1 (note on dialogue teletype waiting)
* Make cel gardens torii light glow signal when passing through the "correct" direction
* Standardize dialogue manager teletype waiting via time vs. by frame
* Change ending label texts
* Add small message on loading screen for unusually long wait times
* Handle Day 2 Woods where Ids runs backward, make run forward and handle if Player is directly in front
* In Ursa Saloon, add dialogue cooldown to bar drinks after Ursie intro
* Specify exactly where Speed Seal should not work. Rather state declaratively than have player check an array, since the check would be every frame
* On Wells World and Mirror Halls 1, do opening dialogue reaction based on Map Notification ending, so the timing is standardized
* Moose should pause after checking for Spell Book
* Underworld totem should keep Player uninteractable for longer to have more weight
* Labyrinth, switch Player to and out of Psychic Duck before/after intro, before/after unique blocking cut scenes (no SFX for switch on since Puppeteer SFX will be occurring), before/after success cut scene (here the switch back will be implied -- without player seeing -- because there's too many cases to handle
* When puppet is blocked, stop its moving animation
* Player should face up after My Mask Mask Effect
* Wait 1 more frame after switching puppeteer off from hold to match timing of puppet deactivation
* Implement equipment UI on enter, where it will take last inventory item (like inventory does)
* Implement done event for default map notification
* Implement fading in logo separately and define separate parameters for initial default load (vs. subsequent loads "FromBack")
* Implement auto-hiding lights to remain under the 8 point light limit
* Fix naming of My Mask in dialogue/UI

Note on teletype dialogue waiting:
If yield return null or wait time is consistently less than frame length (e.g. 0.010f charTypeTime and 30 fps which is .0333 frame length) then each letter will be shown on the next frame available.

If the wait time is greater than the frame length, then it will skip (wait time / frame length) frames in between each letter.

For example, when we set the wait to 0.01111111f, the behavior will be the same as the original code of waiting with yield return null until 100 fps where it will start waiting frames equal to wait time / frame length. This should "normalize" the typing speed in cases where fps is too high like 120 fps.

Attached example shows this behavior pretty clearly. Once the Time.deltaTimes add up to greater than the weight time, the next letter is revealed. So in the case with 120fps, with a wait time of 0.01111111f , (most likely) on frame f1, only ~0.008333 will have passed, so it will skip that frame. Then on f2 ~0.016667 will have passed, so the next letter will be revealed. In this way, at 120 fps, the typing speed should appear how it is at 60 fps, where each letter is shown 1 frame after the next (~0.016667 apart).

https://trello.com/c/i5dTcyPf